+++ Client/src/api/detailPost/Delete/Delete.ts
