+++ Client/src/api/post/getPost.ts
@@ -1,5 +1,4 @@
-import { ArrayCommentType } from "~/@types/detailPost.types";
-import { apiClient } from "~/api/axiosInstance";
+import { apiClient } from "../axiosInstance";
 
 export const getPost = async (
   postId: string | undefined,
@@ -34,15 +33,4 @@ export const getPost = async (
       views: 1,
     };
   }
-};
-
-export const getPostCommentList = async (id: string | undefined) => {
-  let result: ArrayCommentType = [];
-  await apiClient
-    .get(`/comments/listof/${id}`)
-    .then((res) => {
-      result = res.data.data;
-    })
-    .catch((err) => console.error(err));
-  return result;
 };
