+++ Client/src/components/DetailPost/Comment/index.tsx
@@ -2,9 +2,9 @@ import { useState } from "react";
 import { useParams } from "react-router-dom";
 import { useRecoilState, useSetRecoilState } from "recoil";
 import Recomment from "../Recomment";
-import { deletePostComment } from "~/api/BlogDetail/Delete/Delete";
-import { modifiedComment } from "~/api/BlogDetail/Patch/Patch";
-import { handleCommentSubmit } from "~/api/BlogDetail/Post/Post";
+import { deletePostComment } from "~/api/detailPost/Delete/Delete";
+import { modifiedComment } from "~/api/detailPost/Patch/Patch";
+import { handleCommentSubmit } from "~/api/detailPost/Post/Post";
 import { UserDataAtomFamily } from "~/recoil/auth";
 import { isLoginModalVisibleAtom } from "~/recoil/modal/atoms";
 import { IoMdArrowDropup, IoMdArrowDropdown } from "react-icons/io";
