+++ Client/src/pages/DetailPost/index.tsx
@@ -2,9 +2,10 @@ import { useState, useRef, useEffect } from "react";
 import { AiOutlineShareAlt, AiFillEye, AiFillHeart } from "react-icons/ai";
 import { useParams, useNavigate } from "react-router-dom";
 import { useSetRecoilState, useRecoilValue } from "recoil";
-import { deletePostHandler } from "~/api/BlogDetail/Delete/Delete";
-import { getPost, getPostCommentList } from "~/api/BlogDetail/Get/Get";
-import { handleLikePost } from "~/api/BlogDetail/Post/Post";
+import { deletePostHandler } from "~/api/detailPost/Delete/Delete";
+import { getPostCommentList } from "~/api/detailPost/Get/Get";
+import { handleLikePost } from "~/api/detailPost/Post/Post";
+import { getPost } from "~/api/post/getPost";
 import Footer from "~/components/@common/Footer";
 import { DefaultHeader } from "~/components/@common/Header";
 import { AddComment, Comment } from "~/components/DetailPost";
